Support additional lookups, use Launch Services version numbers
Add support for querying the default application, as well as querying
handlers of a given URL.

Replace use of the CFBundleVersion as written with the Launch Services
intepretation of the version number. This requires use of a private API,
but provides a more accurate version comparison for the application's
intended purpose - identifying duplicate bundles with the same version.
Launch Services parses the string and creates its own representation of
the version number that may discard additional information in the
string, so a simple string comparison is inappropriate.